info of thorughtput and CHR has been added

+++ src/main/java/com/github/nkoutroumanis/weatherIntegrator/WeatherIntegrator.java
@@ -146,7 +146,7 @@ public final class WeatherIntegrator {
                 if(numberofRecords%JobFilesUsingIndex.INFOEVERYN == 0){
                     logger.info("CHR {}", ((double) hits / numberofRecords));
                     logger.info("Overall Throughtput {}", ((double) WeatherIntegrator.numberofRecords / ((System.currentTimeMillis() - start) / 1000)));
-                    logger.info("Window Throughtput {}", JobFilesUsingIndex.INFOEVERYN/((System.currentTimeMillis() - startTimeWindow) / 1000));
+                    logger.info("Window Throughtput {}", ((double) JobFilesUsingIndex.INFOEVERYN/((System.currentTimeMillis() - startTimeWindow) / 1000)));
                     logger.info("Opened {}", (numberofRecords - hits) - window);
                     window = numberofRecords - hits;
                     startTimeWindow = System.currentTimeMillis();
